+++ CHANGELOG.md
@@ -10,6 +10,16 @@ internal API changes are not present.
 Main (unreleased)
 -----------------
 
+### Breaking changes
+
+- The algorithm for the "hash" action of `otelcol.processor.attributes` has changed.
+  The change was made in PR [#22831](https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/22831) of opentelemetry-collector-contrib. (@ptodev)
+
+- `otelcol.exporter.loki` now includes the instrumentation scope in its output. (@ptodev)
+
+- `otelcol.extension.jaeger_remote_sampling` removes the `\` HTTP endpoint. The `/sampling` endpoint is still functional.
+  The change was made in PR [#18070](https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/18070) of opentelemetry-collector-contrib. (@ptodev)
+
 ### Features
 
 - The Pyroscope scrape component computes and sends delta profiles automatically when required to reduce bandwidth usage. (@cyriltovena)
@@ -19,7 +29,6 @@ Main (unreleased)
 - Integrations: Introduce the `squid` integration. (@armstrmi)
 
 - Support custom fields in MMDB file for `stage.geoip`. (@akselleirv)
-
 
 - New Grafana Agent Flow components:
 
@@ -65,6 +74,8 @@ Main (unreleased)
 - The Grafana Agent mixin now includes a dashboard for the logs pipeline. (@thampiotr)
 
 - The Agent Operational dashboard of Grafana Agent mixin now has more descriptive panel titles, Y-axis units
+
+- Update OpenTelemetry Collector dependencies from v0.63.0 to v0.79.0. (@ptodev)
 
 ### Bugfixes